+++ internal/ComicRGBAFilter.go
@@ -22,7 +22,7 @@ type ComicRGBAFilter struct {
 	colorStepF             float64
 }
 
-func (filter *ComicRGBA64Filter) Apply(img, filteredImg *image.RGBA64, startY, endY int, prgrsCh chan uint8) {
+func (filter *ComicRGBA64Filter) Apply(img, filteredImg *image.RGBA64, startY, endY int, prgrsCh chan int) {
 	if iter, err := NewImageIterator(img, NONE, startY, endY, prgrsCh); err == nil {
 		for iter.HasNext() {
 			curr := iter.Next()
@@ -34,7 +34,7 @@ func (filter *ComicRGBA64Filter) Apply(img, filteredImg *image.RGBA64, startY, e
 	}
 }
 
-func (filter *ComicRGBAFilter) Apply(img, filteredImg *image.RGBA, startY, endY int, prgrsCh chan uint8) {
+func (filter *ComicRGBAFilter) Apply(img, filteredImg *image.RGBA, startY, endY int, prgrsCh chan int) {
 	if iter, err := NewImageIterator(img, NONE, startY, endY, prgrsCh); err == nil {
 		for iter.HasNext() {
 			curr := iter.Next()
